% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circos_protein_plot.R
\name{circos_protein_plot}
\alias{circos_protein_plot}
\title{Circos Protein Plot}
\usage{
circos_protein_plot(
  circos_data,
  total_track_number,
  track_id_column,
  segment_names_column,
  protein_column,
  beta_column,
  se_column,
  odds_ratios = F,
  custom_palette = viridis::viridis(n = total_track_number + 5)[total_track_number + 5:1],
  primary_track = 1,
  error_bar_ends = TRUE,
  axis_label_size = 0.5,
  text_size = 1.5,
  point_size = 1,
  add_legend = FALSE,
  track_name_column = track_id_column
)
}
\arguments{
\item{circos_data}{data.table containing all of the data to plot}

\item{total_track_number}{total number of tracks to plot (can be less than sources of data if only plotting subset)}

\item{track_id_column}{name of column containing variable to determine which track e.g. method / data source}

\item{segment_names_column}{column containing factor to subset data on}

\item{protein_column}{column containing names of proteins/metabolites}

\item{beta_column}{column containing beta value}

\item{se_column}{column containing standard error}

\item{odds_ratios}{(optional) provide vector of track numbers to exponentiate or T if applying to all tracks (default = F)}

\item{custom_palette}{(optional) provide custom colour palette, supports viridis or custom vector}

\item{primary_track}{variable from "track_id_column" to determine which track is first to plot and used to generate names}

\item{error_bar_ends}{(optional) boolean value, whether to add ends to error bars (default = TRUE)}

\item{axis_label_size}{(optional) custom label size for axis}

\item{text_size}{(optional) custom text size for outer labels}

\item{point_size}{(optional) custom size for points on plots}

\item{add_legend}{add in track legend (default = FALSE)}

\item{track_name_column}{(optional) column containing track names (default = uses track_id_column)}
}
\description{
A solution to displaying multiple sources of data for the same values on multiple tracks with segements based on custom criteria, provided with a data frame, customise which tracks data is displayed on, how many tracks are required, the segments to subset the data by, custom colour scheme, generate odds ratios and deal with missing data. This package uses the 'circlize' package to create beautiful publication ready circular plots for protein or metabolite data from a number of different analyses, adjusting track size for the best visualisation.
}
