% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circos_protein_plot_segment.R
\name{circos_protein_plot_segment}
\alias{circos_protein_plot_segment}
\title{circos protein plot segment}
\usage{
circos_protein_plot_segment(
  circos_data,
  total_track_number,
  track_id_column,
  segment_names_column,
  segment_to_display,
  protein_column,
  beta_column,
  se_column,
  odds_ratios,
  custom_palette,
  primary_track,
  error_bar_ends
)
}
\arguments{
\item{circos_data}{data.table containing all of the data to plot}

\item{total_track_number}{total number of tracks to plot (can be less than sources of data if only plotting subset)}

\item{track_id_column}{name of column containing variable to determine which track e.g. method / data source}

\item{segment_names_column}{column containing factor to subset data on}

\item{segment_to_display}{segment to zoom in on}

\item{protein_column}{column containing names of proteins/metabolites}

\item{beta_column}{column containing beta value}

\item{se_column}{column containing standard error}

\item{odds_ratios}{(optional) boolean value, whether to generate and plot odds ratios from beta and se (default = FALSE)}

\item{custom_palette}{provide custom colour palette, supports viridis or custom vector}

\item{primary_track}{variable from "track_id_column" to determine which track is first to plot and used to generate names}

\item{error_bar_ends}{(optional) boolean value, whether to add ends to error bars (default = TRUE)}
}
\description{
circos protein plot segment
}
